Spacer: Convert knobs to controls in Storybook
