+++ packages/components/src/spacer/component.tsx
@@ -11,7 +11,7 @@ import { View } from '../view';
 import { useSpacer } from './hook';
 import type { Props } from './types';
 
-function Spacer(
+function UnconnectedSpacer(
 	props: WordPressComponentProps< Props, 'div' >,
 	forwardedRef: ForwardedRef< any >
 ) {
@@ -25,7 +25,6 @@ function Spacer(
  *
  * `Spacer` comes with a bunch of shorthand props to adjust `margin` and `padding`. The values of these props work as a multiplier to the library's grid system (base of `4px`).
  *
- * @example
  * ```jsx
  * import { Spacer } from `@wordpress/components`
  *
@@ -43,6 +42,6 @@ function Spacer(
  * }
  * ```
  */
-const ConnectedSpacer = contextConnect( Spacer, 'Spacer' );
+export const Spacer = contextConnect( UnconnectedSpacer, 'Spacer' );
 
-export default ConnectedSpacer;
+export default Spacer;
